refactor: change file structure in folder `./src/main/profile`
