+++ src/view/main/profile/components/password/password-model.ts
@@ -1,7 +1,7 @@
 import { ClientResponse, Customer, CustomerSignInResult } from '@commercetools/platform-sdk';
 import { ChangePasswordDto } from '#src/api/endpoints/types/customer';
 import { setProfile } from '#src/logic/state/state';
-import BaseModel from './base-model';
+import BaseModel from '../base-model';
 
 export default class PasswordModel extends BaseModel {
   protected async apiPromiseChain(data: unknown): Promise<void> {
