+++ src/view/main/profile/components/user-data/personales-model.ts
@@ -1,7 +1,7 @@
 import { ClientResponse, Customer } from '@commercetools/platform-sdk';
 import { PersonalesDto } from '#src/api/endpoints/types/customer';
 import { setProfile } from '#src/logic/state/state';
-import BaseModel from './base-model';
+import BaseModel from '../base-model';
 
 export default class PersonalesModel extends BaseModel {
   protected async apiPromiseChain(data: unknown): Promise<void> {
